move test to a better location
