+++ cloud/filestore/tests/config_dispatcher/ya.make
@@ -10,7 +10,6 @@ PEERDIR(
 
     library/python/testing/yatest_common
 
-    contrib/python/protobuf
     contrib/ydb/tests/library
 
     contrib/python/requests/py3
