+++ cloud/filestore/tests/python/lib/daemon_config/ya.make
@@ -1,0 +1,13 @@
+PY3TEST()
+
+TEST_SRCS(test.py)
+
+PEERDIR(
+    cloud/filestore/config
+    cloud/filestore/tests/python/lib
+
+    contrib/python/protobuf
+    contrib/ydb/core/protos
+)
+
+END()
