+++ cloud/filestore/tests/python/lib/ya.make
@@ -33,3 +33,7 @@ PY_SRCS(
 )
 
 END()
+
+RECURSE_FOR_TESTS(
+    daemon_config
+)
